Set `name` properties for folders and scripts individually
This is part of a bigger refactor of our tree restructure. We're trying to abstract out a lot of things from the main compilier functions into more reusable ones.

We're repeating ourselves currently by setting the names of both these to the basename of the file/folder they're for, but this will look to be changed in  the near future.