Refactor Object hierarchy generation
So for the past 4-5 days I've been working tirelessly to create an assortment of classes and helper functions to have a conrete foundation for creating all of our objects in the tree.

Turns out all of that was way over complicated, and this commit was thrown together in about half an hour. It does everything we need without over-complicating and abstracting our code throughout multiple files.