Move the server package to its own folder
This is in preparation for the plugin we'll need for Roblox Studio. We're going to contain both the server package and plugin in this repo in their own folders.

This commit is to get everything out of the way so we don't have a new 'plugin' folder mixed in with the server code.