Remove refresh rate option
This is better for us to fine-tune behind the scenes. Giving the user complete access to how many HTTP requests are made can become a big problem for the user.

If they set the refresh rate low enough where they were exceeding 500 requests per minute, then Studio Bridge and all other plugins using HttpService would stall for 30 seconds.

In the future we we might add this back, with limits on how low the user can set the refresh rate to prevent the above from happening.