Move the plugin and CLI to their own repositories
Having to fragment each part of the codebase to their own repos isn't too desirable, but in the end it's necessary for the project.

We needed a separate repo for the CLI at the very least, as updating it on NPM requires a bump in version number. Having to share a version number with the plugin wasn't going to work too well with this.

Having to share a commit history between the two projects was proving cumbersome as well. Overall this is positive change for the health of the plugin and CLI.