Remove paragraph (too much info)
This paragraph was breaking the flow of the setup docs by explaining too much about how the CLI works.

The user will likely not be concerned about the inner workings of the CLI. In this early stage of learning about Studio Bridge, we need to present the user with only the information they need to get started.

If they want to learn how the CLI works, they'll go searching for it.